dental_management/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
